feat: migrate app.js, server.js and config.js to TS
